Simplify some of the time handling
This came out of looking at #251. I didn't actually fix that as I fear
it might break people's code, especially tests and such.

So just add a comment to fix it in a v2. And do simplify some things,
and move some of the cutom timestamp parsing to a new internal/pqtime
package, and add a few more tests.
