+++ bench_test.go
@@ -7,7 +7,6 @@ import (
 	"database/sql"
 	"database/sql/driver"
 	"io"
-	"math/rand"
 	"net"
 	"runtime"
 	"strconv"
@@ -17,6 +16,7 @@ import (
 	"time"
 
 	"github.com/lib/pq/internal/pqtest"
+	"github.com/lib/pq/internal/pqtime"
 	"github.com/lib/pq/oid"
 )
 
@@ -364,41 +364,12 @@ func BenchmarkDecodeTimestamptzMultiThread(b *testing.B) {
 	oldProcs := runtime.GOMAXPROCS(0)
 	defer runtime.GOMAXPROCS(oldProcs)
 	runtime.GOMAXPROCS(runtime.NumCPU())
-	globalLocationCache = newLocationCache()
+	pqtime.Reset()
 
 	f := func(wg *sync.WaitGroup, loops int) {
 		defer wg.Done()
 		for i := 0; i < loops; i++ {
 			decode(&parameterStatus{}, testTimestamptzBytes, oid.T_timestamptz, formatText)
-		}
-	}
-
-	wg := &sync.WaitGroup{}
-	b.ResetTimer()
-	for j := 0; j < 10; j++ {
-		wg.Add(1)
-		go f(wg, b.N/10)
-	}
-	wg.Wait()
-}
-
-func BenchmarkLocationCache(b *testing.B) {
-	globalLocationCache = newLocationCache()
-	for i := 0; i < b.N; i++ {
-		globalLocationCache.getLocation(rand.Intn(10000))
-	}
-}
-
-func BenchmarkLocationCacheMultiThread(b *testing.B) {
-	oldProcs := runtime.GOMAXPROCS(0)
-	defer runtime.GOMAXPROCS(oldProcs)
-	runtime.GOMAXPROCS(runtime.NumCPU())
-	globalLocationCache = newLocationCache()
-
-	f := func(wg *sync.WaitGroup, loops int) {
-		defer wg.Done()
-		for i := 0; i < loops; i++ {
-			globalLocationCache.getLocation(rand.Intn(10000))
 		}
 	}
